+++ NEWS.md
@@ -1,3 +1,14 @@
+# Changes in bayesammi Version 0.3.0
+## New Features 
+
+* Added biplot function
+
+# Changes in bayesammi Version 0.2.0
+## New Features 
+
+* Updated the DESCRIPTION file to fulfill the requirements of CRAN
+
+
 # Changes in bayesammi Version 0.1.0
 ## New Features